ISPN-7509 TotalOrderStateTransferInterceptor doesn't handle OutdatedTopologyException for read commands
