+++ core/src/main/java/org/infinispan/interceptors/impl/BaseStateTransferInterceptor.java
@@ -8,11 +8,16 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReferenceFieldUpdater;
 import java.util.function.BiFunction;
 
+import org.infinispan.commands.FlagAffectedCommand;
 import org.infinispan.commands.TopologyAffectedCommand;
 import org.infinispan.commands.VisitableCommand;
+import org.infinispan.commands.read.GetAllCommand;
+import org.infinispan.commands.read.GetCacheEntryCommand;
+import org.infinispan.commands.read.GetKeyValueCommand;
 import org.infinispan.commands.remote.GetKeysInGroupCommand;
 import org.infinispan.configuration.cache.Configuration;
 import org.infinispan.context.InvocationContext;
+import org.infinispan.context.impl.FlagBitSets;
 import org.infinispan.distribution.DistributionManager;
 import org.infinispan.factories.KnownComponentNames;
 import org.infinispan.factories.annotations.ComponentName;
@@ -41,10 +46,11 @@ import org.infinispan.util.logging.LogFactory;
  */
 public abstract class BaseStateTransferInterceptor extends DDAsyncInterceptor {
    private final boolean trace = getLog().isTraceEnabled();
+   private final InvocationFinallyFunction handleReadCommandReturn = this::handleReadCommandReturn;
 
-   protected StateTransferManager stateTransferManager;
+   private StateTransferManager stateTransferManager;
    protected StateTransferLock stateTransferLock;
-   protected Executor remoteExecutor;
+   private Executor remoteExecutor;
    private DistributionManager distributionManager;
    private ScheduledExecutorService timeoutExecutor;
 
@@ -128,7 +134,7 @@ public abstract class BaseStateTransferInterceptor extends DDAsyncInterceptor {
       return cacheTopology == null ? -1 : cacheTopology.getTopologyId();
    }
 
-   protected final void updateTopologyId(TopologyAffectedCommand command) throws InterruptedException {
+   protected final void updateTopologyId(TopologyAffectedCommand command) {
       // set the topology id if it was not set before (ie. this is local command)
       // TODO Make tx commands extend FlagAffectedCommand so we can use CACHE_MODE_LOCAL in TransactionTable.cleanupStaleTransactions
       if (command.getTopologyId() == -1) {
@@ -141,7 +147,7 @@ public abstract class BaseStateTransferInterceptor extends DDAsyncInterceptor {
 
    protected <T extends VisitableCommand> Object retryWhenDone(CompletableFuture<Void> future, int topologyId,
                                                                InvocationContext ctx, T command,
-                                                               InvocationFinallyFunction callback) throws Throwable {
+                                                               InvocationFinallyFunction callback) {
       if (future.isDone()) {
          getLog().tracef("Retrying command %s for topology %d", command, topologyId);
          return invokeNextAndHandle(ctx, command, callback);
@@ -157,6 +163,88 @@ public abstract class BaseStateTransferInterceptor extends DDAsyncInterceptor {
          cancellableRetry.setTimeoutFuture(timeoutFuture);
          return makeStage(asyncInvokeNext(ctx, command, retryFuture)).andHandle(ctx, command, callback);
       }
+   }
+
+   @Override
+   public Object visitGetKeyValueCommand(InvocationContext ctx, GetKeyValueCommand command) throws Throwable {
+      return handleReadCommand(ctx, command);
+   }
+
+   @Override
+   public Object visitGetCacheEntryCommand(InvocationContext ctx, GetCacheEntryCommand command)
+         throws Throwable {
+      return handleReadCommand(ctx, command);
+   }
+
+   @Override
+   public Object visitGetAllCommand(InvocationContext ctx, GetAllCommand command) throws Throwable {
+      return handleReadCommand(ctx, command);
+   }
+
+   protected <C extends VisitableCommand & TopologyAffectedCommand & FlagAffectedCommand> Object handleReadCommand(
+         InvocationContext ctx, C command) {
+      return isLocalOnly(command) ? invokeNext(ctx, command) :
+            updateAndInvokeNextRead(ctx, command);
+   }
+
+   private <C extends VisitableCommand & TopologyAffectedCommand> Object updateAndInvokeNextRead(InvocationContext ctx, C command) {
+      updateTopologyId(command);
+      return invokeNextAndHandle(ctx, command,handleReadCommandReturn);
+   }
+
+   private Object handleReadCommandReturn(InvocationContext rCtx, VisitableCommand rCommand, Object rv, Throwable t)
+         throws Throwable {
+      if (t == null)
+         return rv;
+
+      Throwable ce = t;
+      while (ce instanceof RemoteException) {
+         ce = ce.getCause();
+      }
+      final CacheTopology cacheTopology = stateTransferManager.getCacheTopology();
+      int currentTopologyId = cacheTopology == null ? -1 : cacheTopology.getTopologyId();
+      TopologyAffectedCommand cmd = (TopologyAffectedCommand) rCommand;
+      if (ce instanceof SuspectException) {
+         if (trace)
+            getLog().tracef("Retrying command because of suspected node, current topology is %d: %s",
+                  currentTopologyId, rCommand);
+         // It is possible that current topology is actual but the view still contains a node that's about to leave;
+         // a broadcast to all nodes then can end with suspect exception, but we won't get any new topology.
+         // An example of this situation is when a node sends leave - topology can be installed before the new view.
+         // To prevent suspect exceptions use SYNCHRONOUS_IGNORE_LEAVERS response mode.
+         if (cacheTopology != null && currentTopologyId == cmd.getTopologyId() && !cacheTopology.getActualMembers().contains(((SuspectException) ce).getSuspect())) {
+            // TODO: provide a test case
+            throw new IllegalStateException("Command was not sent with SYNCHRONOUS_IGNORE_LEAVERS?");
+         }
+      } else if (ce instanceof OutdatedTopologyException) {
+         if (trace)
+            getLog().tracef("Retrying command because of topology change, current topology is %d: %s",
+                  currentTopologyId, cmd);
+      } else {
+         throw t;
+      }
+      // We increment the topology to wait for the next topology.
+      // Without this, we could retry the command too fast and we could get the OutdatedTopologyException again.
+      int newTopologyId = getNewTopologyId(ce, currentTopologyId, cmd);
+      cmd.setTopologyId(newTopologyId);
+      ((FlagAffectedCommand)rCommand).addFlags(FlagBitSets.COMMAND_RETRY);
+      CompletableFuture<Void> topologyFuture = stateTransferLock.topologyFuture(newTopologyId);
+      return retryWhenDone(topologyFuture, newTopologyId, rCtx, rCommand, handleReadCommandReturn);
+   }
+
+   protected int getNewTopologyId(Throwable ce, int currentTopologyId, TopologyAffectedCommand command) {
+      int requestedTopologyId = command.getTopologyId() + 1;
+      if (ce instanceof OutdatedTopologyException) {
+         OutdatedTopologyException ote = (OutdatedTopologyException) ce;
+         if (ote.requestedTopologyId >= 0) {
+            requestedTopologyId = ote.requestedTopologyId;
+         }
+      }
+      return Math.max(currentTopologyId, requestedTopologyId);
+   }
+
+   protected boolean isLocalOnly(FlagAffectedCommand command) {
+      return command.hasAnyFlag(FlagBitSets.CACHE_MODE_LOCAL);
    }
 
    protected abstract Log getLog();
@@ -180,7 +268,7 @@ public abstract class BaseStateTransferInterceptor extends DDAsyncInterceptor {
       @SuppressWarnings("unused")
       private volatile Object timeoutFuture;
 
-      public CancellableRetry(T command, int topologyId) {
+      CancellableRetry(T command, int topologyId) {
          this.command = command;
          this.topologyId = topologyId;
       }
